- Se cambia el logo de ITV - Se cambian las fechas y todas las funciones de integer a long pq las fechas futuras son demasiado grandes - Se añaden los procesos de filtro de gasolina, filtro de aire y bujías - Se añaden el resto de variables de kms y fechas máximas de todas las revisiones a tener en cuenta pero sin realizar sus procesos

+++ app/src/main/java/com/carlog/gilberto/carlog/data/DBCar.java
@@ -95,7 +95,7 @@ public class DBCar {
 
     }
 
-    public void ActualizarITVCocheActivo(String matricula, int itv) {
+    public void ActualizarITVCocheActivo(String matricula, long itv) {
         String sql = "UPDATE " + TABLE_NAME + " SET " + CN_ITV + " = '" + itv + "' WHERE " + CN_PROFILE + " = '" + TipoCoche.PROFILE_ACTIVO + "' AND " +CN_MATRICULA + " = '" + matricula + "'";
         db.execSQL(sql);
     }
